+++ src/main/java/com/kirani/kiranibackendcode/repository/OrdersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrdersRepository extends CrudRepository<Orders, Integer> {
+public interface OrdersRepository extends CrudRepository<Orders, Long> {
 
 
 }
